Move RepoData definition into conda.core.repodata module.
